Show Feedback input field only, if it is not your own answer.

+++ app/src/components/OpenFeedback.svelte
@@ -3,6 +3,7 @@
     const dispatch = createEventDispatcher();
     import { DataStore } from "@aws-amplify/datastore";
     import { OpenFeedbackDraft, OpenFeedback, OpenAnswer } from "../models";
+    import { Auth } from "aws-amplify";
 
     export let baseUrl;
     export let openAnswer: OpenAnswer;
@@ -85,44 +86,58 @@
             .then((result) => console.log(result))
             .catch((error) => console.log("error", error));
     }
+
+    async function isOpenAnswerUserOwned(openAnswer: OpenAnswer) {
+        const user = await Auth.currentAuthenticatedUser();
+        const userOwned = openAnswer.owner == user.attributes.sub;
+        console.log(userOwned)
+        return openAnswer.owner == user.attributes.sub;
+    }
 </script>
 
 <div>
     {#if openFeedback}
         Feedback: {openFeedback.feedbackText}
     {:else}
-        {#if openFeedbackDraft}
-            <div class="flex justify-between">
-                <div>{openFeedbackDraft.feedbackText}</div>
+        {#await isOpenAnswerUserOwned(openAnswer) then isUserOwned}
+            {#if !isUserOwned}
+                {#if openFeedbackDraft}
+                    <div class="flex justify-between">
+                        <div>{openFeedbackDraft.feedbackText}</div>
+                        <div>
+                            <button
+                                on:click={() =>
+                                    deleteMyFeedbackDraft(
+                                        openFeedbackDraft,
+                                        openAnswer
+                                    )}
+                                class="w-32">Delete</button
+                            >
+                        </div>
+                    </div>
+                {:else}
+                    <input
+                        class="w-full"
+                        placeholder="Provide feedback to this answer"
+                        on:keydown={(e) => {
+                            if (e.key === "Enter") {
+                                saveOpenFeedbackDraft(
+                                    openAnswer,
+                                    e.target.value
+                                );
+                            }
+                        }}
+                    />
+                {/if}
                 <div>
                     <button
+                        disabled={!openFeedbackDraft}
                         on:click={() =>
-                            deleteMyFeedbackDraft(
-                                openFeedbackDraft,
-                                openAnswer
-                            )}
-                        class="w-32">Delete</button
+                            commitOpenFeedback(openFeedbackDraft, openAnswer)}
+                        class="w-32">Commit</button
                     >
                 </div>
-            </div>
-        {:else}
-            <input
-                class="w-full"
-                placeholder="Provide feedback to this answer"
-                on:keydown={(e) => {
-                    if (e.key === "Enter") {
-                        saveOpenFeedbackDraft(openAnswer, e.target.value);
-                    }
-                }}
-            />
-        {/if}
-        <div>
-            <button
-                disabled={!openFeedbackDraft}
-                on:click={() =>
-                    commitOpenFeedback(openFeedbackDraft, openAnswer)}
-                class="w-32">Commit</button
-            >
-        </div>
+            {/if}
+        {/await}
     {/if}
 </div>
